* Make PlaybackManager::CurrentFrame() return int64, like    what it's internally using. Adapt NotifyCurrentFrameChanged()    implementors.  * Controller::SetTo() does not need to set the current frame,    since either Init() or FormatChanged() will have taken care    of it now.  * Reset the seeking request info in Controller::SetTo().  * Changed the parameter passed to VideoSupplier::FillBuffer()    from media_format to media_raw_video_format, so the    VideoProducer doesn't have to generate a media_format for    each frame...  * ProxyVideoSupplier caches the last produced frame, which    avoids a situation that the VideoProducer asks to generate    the same frame twice sometimes.  * In PlaybackManager::_PushState(), make sure we really    schedule the next current_frame if asked. This avoids    one situation in which the VideoSupplier was asked to generate    the same frame again and fixes playback after pausing (was    showing black video until the next keyframe before.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@38553 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/mediaplayer/Controller.h
@@ -168,7 +168,7 @@ private:
 									bool enabled) const;
 	virtual	void				NotifyVideoBoundsChanged(BRect bounds) const;
 	virtual	void				NotifyFPSChanged(float fps) const;
-	virtual	void				NotifyCurrentFrameChanged(int32 frame) const;
+	virtual	void				NotifyCurrentFrameChanged(int64 frame) const;
 	virtual	void				NotifySpeedChanged(float speed) const;
 	virtual	void				NotifyFrameDropped() const;
 	virtual	void				NotifyStopFrameReached() const;
